backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
pandas>=2.2.0
datasets>=2.16.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
numpy>=1.26.0